requirements.txt
django==4.2.4
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.2
djangorestframework-simplejwt==5.3.0